Block digging of containers holding any objects
If a space is freed without objects being removed, they might show up
later in a new container.

This way isn't ideal since it relies on continuously checking active
insides. It won't always catch all the entities. However, it should work
almost all the time, and the results of failure are not terrible. I
can't find a way to just detect static objects.