Prevent items passing through disconnected ports
This makes the behavior like that of the filter injector of
Pipeworks. This also fixes an issue with an object not being
counted: Before, if you sent an item into an empty container
through a tube then broke the container, you would succeed, since
the item object wouldn't be counted. Now, that situation is just
impossible